+++ lib/src/scope_blue_print.dart
@@ -342,13 +342,13 @@ class ScopeBluePrint {
   // ...........................................................................
   /// Turns the blue print into a scope and adds it to the parent scope.
   Scope instantiate({
-    required Scope hostScope,
+    required Scope scope,
     Map<String, String> connect = const {},
     bool initScBuilders = true,
     Owner<Scope>? owner,
   }) {
     // Smart scopes must not be instantiated in smart scopes
-    if (isSmartScope && hostScope.isSmartScope) {
+    if (isSmartScope && scope.isSmartScope) {
       throw ArgumentError(
         'Smart scopes must not be instantiated in smart scopes.',
       );
@@ -362,7 +362,7 @@ class ScopeBluePrint {
     final self = _applyConnections(this, {...connections});
 
     // Create an inner scope
-    final innerScope = Scope(parent: hostScope, bluePrint: self, owner: owner);
+    final innerScope = Scope(parent: scope, bluePrint: self, owner: owner);
 
     // Make sure there are no duplicate keys
     _checkForDuplicateKeys(self.nodes);
@@ -378,7 +378,7 @@ class ScopeBluePrint {
     // Init sub scopes
     for (final child in self.children) {
       child.instantiate(
-        hostScope: innerScope,
+        scope: innerScope,
 
         /// ScBuilders are initialized after all nodes are created
         initScBuilders: false,